Removed old migration code that resets cache folders. Alpha version app reset code is moved to alpha build flavour folder. BuildConfig.ALPHA is removed.
