+++ app/src/release/java/org/gdg/frisbee/android/app/BaseApp.java
@@ -3,4 +3,7 @@ package org.gdg.frisbee.android.app;
 import android.app.Application;
 
 public class BaseApp extends Application {
+
+    protected void onAppUpdate(int oldVersion, int newVersion) {
+    }
 }
